+++ core/src/trezor/ui/components/tt/pin.py
@@ -41,16 +41,40 @@ class PinInput(ui.Component):
             self.repaint = False
 
     def render_pin(self) -> None:
-        display.bar(0, 0, ui.WIDTH, 50, ui.BG)
-        count = len(self.pin)
+        MAX_LENGTH = const(14)  # maximum length of displayed PIN
+        CONTD_MARK = "<"
         BOX_WIDTH = const(240)
         DOT_SIZE = const(10)
-        PADDING = const(14)
+        PADDING = const(4)
         RENDER_Y = const(20)
-        render_x = (BOX_WIDTH - count * PADDING) // 2
+        TWITCH = const(3)
+
+        display.bar(0, 0, ui.WIDTH, 50, ui.BG)
+
+        if len(self.pin) > MAX_LENGTH:
+            contd_width = display.text_width(CONTD_MARK, ui.BOLD) + PADDING
+            twitch = TWITCH * (len(self.pin) % 2)
+        else:
+            contd_width = 0
+            twitch = 0
+
+        count = min(len(self.pin), MAX_LENGTH)
+        render_x = (BOX_WIDTH - count * (DOT_SIZE + PADDING) - contd_width) // 2
+
+        if contd_width:
+            display.text(
+                render_x, RENDER_Y + DOT_SIZE, CONTD_MARK, ui.BOLD, ui.GREY, ui.BG
+            )
+
         for i in range(0, count):
             display.bar_radius(
-                render_x + i * PADDING, RENDER_Y, DOT_SIZE, DOT_SIZE, ui.GREY, ui.BG, 4
+                render_x + contd_width + twitch + i * (DOT_SIZE + PADDING),
+                RENDER_Y,
+                DOT_SIZE,
+                DOT_SIZE,
+                ui.GREY,
+                ui.BG,
+                4,
             )
 
     def render_prompt(self) -> None:
@@ -82,7 +106,7 @@ class PinDialog(ui.Layout):
         prompt: str,
         subprompt: Optional[str],
         allow_cancel: bool = True,
-        maxlength: int = 9,
+        maxlength: Optional[int] = None,
     ) -> None:
         self.maxlength = maxlength
         self.input = PinInput(prompt, subprompt, "")
@@ -121,10 +145,10 @@ class PinDialog(ui.Layout):
             btn.dispatch(event, x, y)
 
     def assign(self, pin: str) -> None:
-        if len(pin) > self.maxlength:
+        if self.maxlength is not None and len(pin) > self.maxlength:
             return
         for btn in self.pin_buttons:
-            if len(pin) < self.maxlength:
+            if self.maxlength is None or len(pin) < self.maxlength:
                 btn.enable()
             else:
                 btn.disable()