+++ core/tests/test_trezor.config.py
@@ -1,7 +1,6 @@
 from common import *
 
 from trezor.crypto import random
-from trezor.pin import pin_to_int
 
 from trezor import config
 
@@ -27,7 +26,7 @@ class TestConfig(unittest.TestCase):
     def test_wipe(self):
         config.init()
         config.wipe()
-        self.assertEqual(config.unlock(pin_to_int(''), None), True)
+        self.assertEqual(config.unlock('', None), True)
         config.set(1, 1, b'hello')
         config.set(1, 2, b'world')
         v0 = config.get(1, 1)
@@ -44,7 +43,7 @@ class TestConfig(unittest.TestCase):
         for _ in range(128):
             config.init()
             config.wipe()
-            self.assertEqual(config.unlock(pin_to_int(''), None), True)
+            self.assertEqual(config.unlock('', None), True)
             appid, key = random_entry()
             value = random.bytes(16)
             config.set(appid, key, value)
@@ -58,7 +57,7 @@ class TestConfig(unittest.TestCase):
     def test_public(self):
         config.init()
         config.wipe()
-        self.assertEqual(config.unlock(pin_to_int(''), None), True)
+        self.assertEqual(config.unlock('', None), True)
 
         appid, key = random_entry()
 
@@ -82,21 +81,30 @@ class TestConfig(unittest.TestCase):
         self.assertEqual(v2, value16)
 
     def test_change_pin(self):
+        PIN1 = '000'
+        PIN2 = '3141592653589793238462643383279502884197'
         config.init()
         config.wipe()
-        self.assertEqual(config.unlock(pin_to_int(''), None), True)
+        self.assertTrue(config.unlock('', None))
         with self.assertRaises(RuntimeError):
             config.set(PINAPP, PINKEY, b'value')
-        self.assertEqual(config.change_pin(pin_to_int('000'), pin_to_int('666'), None, None), False)
-        self.assertEqual(config.change_pin(pin_to_int(''), pin_to_int('000'), None, None), True)
+        self.assertFalse(config.change_pin(PIN1, '666', None, None))
+        self.assertTrue(config.change_pin('', PIN1, None, None))
         self.assertEqual(config.get(PINAPP, PINKEY), None)
         config.set(1, 1, b'value')
+        config.change_pin(PIN1, PIN2, None, None)
         config.init()
-        self.assertEqual(config.unlock(pin_to_int('000'), None), True)
-        config.change_pin(pin_to_int('000'), pin_to_int(''), None, None)
+        self.assertFalse(config.unlock(PIN1, None))
+        with self.assertRaises(RuntimeError):
+            config.set(PINAPP, PINKEY, b'value')
+        self.assertTrue(config.unlock(PIN2, None))
+        self.assertEqual(config.get(1, 1), b'value')
         config.init()
-        self.assertEqual(config.unlock(pin_to_int('000'), None), False)
-        self.assertEqual(config.unlock(pin_to_int(''), None), True)
+        self.assertTrue(config.unlock(PIN2, None))
+        config.change_pin(PIN2, '', None, None)
+        config.init()
+        self.assertFalse(config.unlock(PIN2, None))
+        self.assertTrue(config.unlock('', None))
         self.assertEqual(config.get(1, 1), b'value')
 
     def test_change_sd_salt(self):
@@ -106,37 +114,37 @@ class TestConfig(unittest.TestCase):
         # Enable PIN and SD salt.
         config.init()
         config.wipe()
-        self.assertTrue(config.unlock(pin_to_int(''), None))
+        self.assertTrue(config.unlock('', None))
         config.set(1, 1, b'value')
-        self.assertFalse(config.change_pin(pin_to_int(''), pin_to_int(''), salt1, None))
-        self.assertTrue(config.change_pin(pin_to_int(''), pin_to_int('000'), None, salt1))
+        self.assertFalse(config.change_pin('', '', salt1, None))
+        self.assertTrue(config.change_pin('', '000', None, salt1))
         self.assertEqual(config.get(1, 1), b'value')
 
         # Disable PIN and change SD salt.
         config.init()
-        self.assertFalse(config.unlock(pin_to_int('000'), None))
+        self.assertFalse(config.unlock('000', None))
         self.assertIsNone(config.get(1, 1))
-        self.assertTrue(config.unlock(pin_to_int('000'), salt1))
-        self.assertTrue(config.change_pin(pin_to_int('000'), pin_to_int(''), salt1, salt2))
+        self.assertTrue(config.unlock('000', salt1))
+        self.assertTrue(config.change_pin('000', '', salt1, salt2))
         self.assertEqual(config.get(1, 1), b'value')
 
         # Disable SD salt.
         config.init()
-        self.assertFalse(config.unlock(pin_to_int('000'), salt2))
+        self.assertFalse(config.unlock('000', salt2))
         self.assertIsNone(config.get(1, 1))
-        self.assertTrue(config.unlock(pin_to_int(''), salt2))
-        self.assertTrue(config.change_pin(pin_to_int(''), pin_to_int(''), salt2, None))
+        self.assertTrue(config.unlock('', salt2))
+        self.assertTrue(config.change_pin('', '', salt2, None))
         self.assertEqual(config.get(1, 1), b'value')
 
         # Check that PIN and SD salt are disabled.
         config.init()
-        self.assertTrue(config.unlock(pin_to_int(''), None))
+        self.assertTrue(config.unlock('', None))
         self.assertEqual(config.get(1, 1), b'value')
 
     def test_set_get(self):
         config.init()
         config.wipe()
-        self.assertEqual(config.unlock(pin_to_int(''), None), True)
+        self.assertEqual(config.unlock('', None), True)
         for _ in range(32):
             appid, key = random_entry()
             value = random.bytes(128)
@@ -147,7 +155,7 @@ class TestConfig(unittest.TestCase):
     def test_compact(self):
         config.init()
         config.wipe()
-        self.assertEqual(config.unlock(pin_to_int(''), None), True)
+        self.assertEqual(config.unlock('', None), True)
         appid, key = 1, 1
         for _ in range(259):
             value = random.bytes(259)
@@ -158,7 +166,7 @@ class TestConfig(unittest.TestCase):
     def test_get_default(self):
         config.init()
         config.wipe()
-        self.assertEqual(config.unlock(pin_to_int(''), None), True)
+        self.assertEqual(config.unlock('', None), True)
         for _ in range(128):
             appid, key = random_entry()
             value = config.get(appid, key)